fix(InternalTextLink): Fix for scrollable elements
The existing code made an assumption the entire page was being scrolled
when calculating the scroll destination. However, in DEMI the content is
in a separate scrollable `<div>` while the `<body>` remains scrolled to
the top. When our smooth scrolling code ran it ended up trying to scroll
the body down which ended up being a no-op.

I think the intent of the original override was to prevent the header
text from being cut in half. We can kind of hack this by instead
scrolling our *parent* in to view, which happens to be a stack div.