feat(InternalLink): Add component
We've grown a few of these style of link in our internal codebases.
Unfortunately they generally do not retain the same smarts as e.g.
SmartTextLink, and so the `v` query parameter is often lost. This aims
to be a general purpose Link that isn't bound to Braid's Text rules.