fix(List): Improve markup semantics
The CSS Grid implementation was nice to work with, but it was gross in
that it weaved in `<span>`s:

```html
<ul>
  <span>bullet or number here</span>
  <li>actual list item here</li>
</ul>
```

This is a shot at a CSS Table implementation where the `<li>` can be
modelled as a table row and contain both the bullet/number and content
as cells.

The blocker currently is that I can't figure out how to elegantly stop
the table from overflowing its parent. An aggressive `overflow-wrap`
gets us most of the way there, but it means that table overflow
behaviour is different inside and outside of a list, and it doesn't work
at all for `CodeBlock`s.