feat(MdxProvider): Add link title support
This supports the optional title on Markdown links, e.g.

```
My favorite search engine is [Duck Duck Go](https://duckduckgo.com "The best search engine for privacy").
```

This handles the title directly in `useMdxComponents` instead of adding
support to `<SmartTextLink>`. It seems more Braid-y to make the consumer
compose the component with a `Box` to get a title.