feat(remark): Merge adjacent code blocks (#94)
A new `mergeCodeBlocks` Remark plugin aggregates adjacent code blocks
into a single node's `meta` field for MDX to pick up and map to a
CodeBlock.

The CodeBlock component also gets a bit smarter. It tries to infer
labels based on language combinations (e.g. two HTTPs means request and
response) and ensures label uniqueness.

This wreaks havoc on our visual snapshots, so I deployed a bunch of
tactical HRs to disrupt the aggregation. This led to styling of said
HRs.

Co-authored-by: Ryan Cumming <etaoins@gmail.com>