fix(svg): Strip mermaid inline styles only
mermaid generates some stubborn inline styles that aren't easy to
override with our custom `style.css`. We can strip these from the markup
using SVGO's `removeAttrs` plugin, as seen in #304.

That change added the plugin to our Webpack loaders, which was
configured to run across all SVGs. While this had a slight benefit to
debuggability (we could observe the effect of `removeAttrs` through the
`svgo-loader` alone without having to regenerate the source mermaid
files), it also meant that the plugin was applied to all `.svg`s.

It makes more sense to run this transform in our mermaid plugin so it
only applies to mermaid diagrams. We retain an un-SVGOed copy to
preserve some debuggability.