feat(webpack): Load images in Storybook
I was stuck for a while on why internal images weren't loaded in our
Storybook. Turns out this was an obvious issue with our Webpack
configuration.

This presented an opportunity to finally get some mermaid snapshots in
for regression testing. Storybook is now a pretty good local dev
experience for tweaking mermaid styles; previously, this had to be
tediously tested out on consumer repos. It does require a change to
`style.css` and a change to the mermaid diagram itself to force a
rerender, but is otherwise rather convenient.

A couple oversights in our mermaid styles have been corrected:

- Flowchart: circle nodes are now coloured correctly
- Sequence: activations and notes are now coloured correctly