feat(InternalLink): Support passing history state
This allows passing internal state when a link is followed.

An example where this would be useful is our Developer Site. We could be
aggressive in scrolling to the active menu item on page navigation, but
it would be disruptive to do that when the navigation was started from
the menu itself. We could use this to pass state to suppress scrolling
for the `InternalLink`s in the menu.