fix: Switch Webpack plugins back to JavaScript
This reverts most of #354. We're included cross-package from other Sku
config so we knew this would break if included from an older version of
Sku that didn't support TypeScript configs.

However, it seems this doesn't work even from TypeScript Sku configs;
our Webpack plugins get incorrectly parsed as JavaScript.

It's possible this would work if we distributed JavaScript but we're a
`compilePackage`.