mermaid: Support `prefers-color-scheme`
I'm not sure that this is immediately viable. It looks quite nice when
using an extension like Dark Reader, but without it, the lack of
built-in support for dark mode on `developer.seek.com` means that the
colouring looks rather out of place.

A stopgap may be to try to customise the generation pipeline to add a
background colour to the SVGs, but that would still look like a giant
black blob on an otherwise light page out of the box.